canonicalize album paths
Introduce requirement that album paths always end with a trailing slash.
If an album is read from the config file without a trailing slash,
silently add it.

If a request comes in and the path does not have a trailing slash but
would match an album if it had one, issue a redirect to the album's
canonical resource.