add Slug method to Renderable interface
In order to add individual image pages with permalinks we're going to
want Photos to be able to render their slugs. For the slug I decided to
keep things simple and just use the object's key minus the album's
prefix.

Because of the way listing objects from S3 works, an image is
guaranteed not to have slashes in it after the album's prefix, so we can
get the slug by simply splitting the key on "/" and taking only the last
part.